+++ tests/testthat/test-plotSPC.R
@@ -26,15 +26,18 @@ test_that("plotSPC: aqp.env settings", {
   explainPlotSPC(sp1)
 
   # get plotting details from aqp environment
-  lsp <- get('last_spc_plot', envir=aqp.env)
+  lsp <- get('last_spc_plot', envir = aqp.env)
 
   # should be a list
   expect_true(is.list(lsp))
 
   # check for required components
-  expect_equal(names(lsp), c("width", "plot.order", "x0", "pIDs", "idname", "y.offset", "scaling.factor",
-                             "max.depth", "n", "extra_x_space", "extra_y_space"))
-
+  expect_true(
+    all(
+      c("width", "plot.order", "x0", "pIDs", "idname", "y.offset", "scaling.factor","max.depth", "n", "extra_x_space", "extra_y_space", "hz.depth.LAI") %in% names(lsp)
+    )
+  )
+    
   # basic integrity checks
   expect_equal(profile_id(sp1), lsp$pIDs)
   expect_equal(idname(sp1), lsp$idname)